update Sampling1D
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/sgdm@2550 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/datadriven/basis/linear/noboundary/operation/OperationDensityMargTo1DLinear.cpp
@@ -32,7 +32,7 @@ namespace datadriven
 
 	void OperationDensityMargTo1DLinear::marg_next_dim(base::Grid* g_in, base::DataVector* a_in, base::Grid* &g_out, base::DataVector* &a_out, size_t dims, size_t dim_x, size_t &count) {
 
-		size_t op_dim = (count < dim_x)? 0 : 1;
+		unsigned int op_dim = (count < dim_x)? 0 : 1;
 
 		base::Grid* g_tmp = NULL;
 		base::DataVector* a_tmp = new base::DataVector(1);
